refactor: move requestTokens from walletConnect saga to tokens saga
